pyproject.toml
[project]
name = "notepm-mcp-server"
version = "0.1.0"
description = "A NotePM MCP server implementation."
readme = "README.md"
authors = [{ name = "Keita Ibaraki", email = "ibaraki@ncdc.co.jp" }]
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "mypy>=1.15.0",
    "python-dotenv>=1.1.0",
]

[project.scripts]
notepm-mcp-server = "notepm_mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

